+++ src/ctds/include/tds.h
@@ -52,6 +52,10 @@ enum ParamStyle {
 #  define CTDS_HAVE_READONLY_INTENT 1
 #endif
 
+#if defined(DBSETNTLMV2)
+#  define CTDS_HAVE_NTLMV2 1
+#endif
+
 /*
     Use `sp_executesql` when possible for the execute*() methods. This method
     won't work on older versions of FreeTDS which don't properly support passing